api/requirements.txt
flask
python-dotenv
pymongo
bcrypt
pycryptodome
requests
pyjwt
dnspython>=2.6.1 # not directly required, pinned by Snyk to avoid a vulnerability
werkzeug>=2.3.8 # not directly required, pinned by Snyk to avoid a vulnerability
